뉴스 감정 분석 업그레이드

+++ src/main/java/org/example/stockdashboard/model/dto/SentimentAnalysisResult.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 
 public record SentimentAnalysisResult(
         LocalDateTime date,
-        double positive,
-        double negative,
-        double neutral
+        int positive,
+        int negative,
+        int neutral
 ) {
 }
